+++ src/main/java/com/nnk/springboot/dto/NewCurvePointDTO.java
@@ -1,6 +1,7 @@
 package com.nnk.springboot.dto;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import javax.validation.constraints.Digits;
 import javax.validation.constraints.NotNull;
@@ -32,7 +33,7 @@ public class NewCurvePointDTO implements Serializable {
     @NotNull(message = "Term is mandatory.")
 	@Positive(message = "Term must be strictly positive.")
 	@Digits(integer = 10, fraction = 2, message = "Term must be a decimal number.")
-	private double term;
+	private BigDecimal term;
 	
 	/**
 	 * Value.
@@ -40,5 +41,5 @@ public class NewCurvePointDTO implements Serializable {
     @NotNull(message = "Value is mandatory.")
 	@Positive(message = "Value must be strictly positive.")
 	@Digits(integer = 10, fraction = 2, message = "Value must be a decimal number.")
-	private double value;
+	private BigDecimal value;
 }
